# target/classes/openapi.yaml
openapi: 3.0.1
info:
  description: "**TMF API Reference : TMF - 629 Customer Management**"
  title: Customer Management
  version: 5.0.1
servers:
- url: "{apiRoot}/tmf-api/customer/v5/"
  variables:
    apiRoot:
      default: https://serverRoot
      enum:
      - https://serverRoot
...
paths:
  /customer:
    get:
      description: List or find Customer objects
      operationId: listCustomer
      parameters:
      - description: Comma-separated properties to be provided in response
        explode: true
        in: query
        name: fields
        required: false
        schema:
          type: string
        style: form
      - description: Requested index for start of resources to be provided in response
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
      - description: Requested number of resources to be provided in response
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Customer_list_example:
                  $ref: '#/components/examples/Customer_list_example_response'
              schema:
                items:
                  $ref: '#/components/schemas/Customer'
                type: array
          description: Success
          headers:
            X-Total-Count:
              $ref: '#/components/headers/X-Total-Count'
            X-Result-Count:
              $ref: '#/components/headers/X-Result-Count'
        "400":
          content:
            application/json:
              examples:
                basic:
                  summary: Bad Request Error
                  value:
                    '@type': Error
                    code: ERR400
                    reason: Bad Request
                    message: The server could not understand the request due to invalid
                      syntax. Please correct the syntax and try again.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              examples:
                basic:
                  summary: Unauthorized Error
                  value:
                    '@type': Error
                    code: ERR401
                    reason: Unauthorized
                    message: You are not authorized to access the requested resource.
                      Please provide the correct credentials.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                basic:
                  summary: Forbidden Error
                  value:
                    '@type': Error
                    code: ERR403
                    reason: Forbidden
                    message: You don't have the permission to access the requested
                      resource. It is either read-protected or not readable by the
                      server.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                basic:
                  summary: Not Found Error
                  value:
                    '@type': Error
                    code: ERR404
                    reason: Not Found
                    message: The requested URL was not found on the server.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json:
              examples:
                basic:
                  summary: Method Not allowed Error
                  value:
                    '@type': Error
                    code: ERR405
                    reason: Method Not allowed
                    message: The method is not allowed for the requested URL.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "500":
          content:
            application/json:
              examples:
                basic:
                  summary: Internal Server Error example
                  value:
                    '@type': Error
                    code: ERR500
                    reason: Internal Server Error
                    message: The server encountered an internal error. Please contact
                      the system administrator.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
        "501":
          content:
            application/json:
              examples:
                basic:
                  summary: Not Implemented Error
                  value:
                    '@type': Error
                    code: ERR501
                    reason: The requested operation is not implemented
                    message: The requested operation is not implemented. Please contact
                      the system administrator.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Implemented
        "503":
          content:
            application/json:
              examples:
                basic:
                  summary: Service Unavailable Error
                  value:
                    '@type': Error
                    code: ERR503
                    reason: Service Unavailable
                    message: The service is currently unavailable. Please try again
                      later.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
      summary: List or find Customer objects
      tags:
      - customer
      x-accepts:
      - application/json
      x-tags:
      - tag: customer
    post:
      description: This operation creates a Customer entity.
      operationId: createCustomer
      parameters:
      - description: Comma-separated properties to be provided in response
        explode: true
        in: query
        name: fields
        required: false
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/Customer_FVO'
      responses:
        "201":
          content:
            application/json:
              examples:
                Customer_Create_example:
                  $ref: '#/components/examples/Customer_Create_example_response'
              schema:
                $ref: '#/components/schemas/Customer'
          description: OK/Created
        "202":
          description: Accepted
        "400":
          content:
            application/json:
              examples:
                basic:
                  summary: Bad Request Error
                  value:
                    '@type': Error
                    code: ERR400
                    reason: Bad Request
                    message: The server could not understand the request due to invalid
                      syntax. Please correct the syntax and try again.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              examples:
                basic:
                  summary: Unauthorized Error
                  value:
                    '@type': Error
                    code: ERR401
                    reason: Unauthorized
                    message: You are not authorized to access the requested resource.
                      Please provide the correct credentials.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                basic:
                  summary: Forbidden Error
                  value:
                    '@type': Error
                    code: ERR403
                    reason: Forbidden
                    message: You don't have the permission to access the requested
                      resource. It is either read-protected or not readable by the
                      server.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                basic:
                  summary: Not Found Error
                  value:
                    '@type': Error
                    code: ERR404
                    reason: Not Found
                    message: The requested URL was not found on the server.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json:
              examples:
                basic:
                  summary: Method Not allowed Error
                  value:
                    '@type': Error
                    code: ERR405
                    reason: Method Not allowed
                    message: The method is not allowed for the requested URL.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json:
              examples:
                basic:
                  summary: Conflict Error
                  value:
                    '@type': Error
                    code: ERR409
                    reason: Conflict
                    message: The request could not be completed due to a conflict
                      with the current state of the target resource.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json:
              examples:
                basic:
                  summary: Internal Server Error example
                  value:
                    '@type': Error
                    code: ERR500
                    reason: Internal Server Error
                    message: The server encountered an internal error. Please contact
                      the system administrator.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
        "501":
          content:
            application/json:
              examples:
                basic:
                  summary: Not Implemented Error
                  value:
                    '@type': Error
                    code: ERR501
                    reason: The requested operation is not implemented
                    message: The requested operation is not implemented. Please contact
                      the system administrator.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Implemented
        "503":
          content:
            application/json:
              examples:
                basic:
                  summary: Service Unavailable Error
                  value:
                    '@type': Error
                    code: ERR503
                    reason: Service Unavailable
                    message: The service is currently unavailable. Please try again
                      later.
                    referenceError: https://host/errors
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
      summary: Creates a Customer
      tags:
      - customer
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: customer
...
components:
  examples:
    Customer_list_example_response:
      description: Here is an example of a response for retrieving multiple Customers.
      value:
      - '@type': Customer
        href: https://host:port/tmf-api/customerManagement/v5/customer/1140
        id: "1140"
        name: Moon Football Club
        status: Created
        engagedParty:
          '@type': PartyRef
          href: https://host:port/tmf-api/partyManagement/v5/organization/500
          id: "500"
          name: Happy Travellers
          '@referredType': Organization
        contactMedium:
        - '@type': PhoneContactMedium
          preferred: true
          contactType: homePhone
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          phoneNumber: 01 09 75 83 51
        - '@type': GeographicAddressContactMedium
          preferred: false
          contactType: homeAddress
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          city: Paris
          country: France
          postCode: "75014"
          street1: 15 Rue des Canards
        relatedParty:
        - '@type': RelatedPartyRefOrPartyRoleRef
          role: bill receiver
          partyOrPartyRole:
            '@type': PartyRef
            href: https://host:port/tmf-api/partyManagement/v5/organization/2777
            id: "2777"
            name: John Doe (Accounting) Ltd
            '@referredType': Organization
...
    Customer_Create_example_request:
      description: Here is an example of a request for creating a customer.
      value:
        '@type': Customer
        name: Moon Football Club
        engagedParty:
          '@type': PartyRef
          href: https://host:port/tmf-api/partyManagement/v5/organization/500
          id: "500"
          name: Happy Travellers
          '@referredType': Organization
        contactMedium:
        - '@type': PhoneContactMedium
          preferred: true
          contactType: homePhone
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          phoneNumber: 01 09 75 83 51
        - '@type': GeographicAddressContactMedium
          preferred: false
          contactType: homeAddress
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          city: Paris
          country: France
          postCode: "75014"
          street1: 15 Rue des Canards
        relatedParty:
        - '@type': RelatedPartyRefOrPartyRoleRef
          role: bill receiver
          partyOrPartyRole:
            '@type': PartyRef
            href: https://host:port/tmf-api/partyManagement/v5/organization/2777
            id: "2777"
            name: John Doe (Accounting) Ltd
            '@referredType': Organization
    Customer_Create_example_response:
      description: Here is an example of a response for creating a customer.
      value:
        '@type': Customer
        href: https://host:port/tmf-api/customerManagement/v5/customer/1140
        id: "1140"
        name: Moon Football Club
        status: Created
        engagedParty:
          '@type': PartyRef
          href: https://host:port/tmf-api/partyManagement/v5/organization/500
          id: "500"
          name: Happy Travellers
          '@referredType': Organization
        contactMedium:
        - '@type': PhoneContactMedium
          preferred: true
          contactType: homePhone
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          phoneNumber: 01 09 75 83 51
        - '@type': GeographicAddressContactMedium
          preferred: false
          contactType: homeAddress
          validFor:
            startDateTime: 2018-06-12T00:00:00Z
            endDateTime: 2019-01-01T00:00:00Z
          city: Paris
          country: France
          postCode: "75014"
          street1: 15 Rue des Canards
        relatedParty:
        - '@type': RelatedPartyRefOrPartyRoleRef
          role: bill receiver
          partyOrPartyRole:
            '@type': PartyRef
            href: https://host:port/tmf-api/partyManagement/v5/organization/2777
            id: "2777"
            name: John Doe (Accounting) Ltd
            '@referredType': Organization
...
  headers:
    X-Total-Count:
      description: Total number of items matching criteria
      explode: false
      schema:
        type: integer
      style: simple
    X-Result-Count:
      description: Actual number of items returned in the response body
      explode: false
      schema:
        type: integer
      style: simple
...
  requestBodies:
    Customer_FVO:
      content:
        application/json:
          examples:
            Customer_Create_example:
              $ref: '#/components/examples/Customer_Create_example_request'
          schema:
            $ref: '#/components/schemas/Customer_FVO'
      description: The Customer to be created
      required: true
...
  schemas:
    AccountRef:
      allOf:
      - $ref: '#/components/schemas/EntityRef'
      description: Account reference. A account may be a party account or a financial
        account.
      discriminator:
        mapping:
          AccountRef: '#/components/schemas/AccountRef'
        propertyName: '@type'
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        href: href
        id: id
        '@schemaLocation': '@schemaLocation'
      type: object
...
    BusinessPartner:
      allOf:
      - $ref: '#/components/schemas/PartyRole'
      description: 'When business partner is the BusinessPartner '
      type: object
...
    Customer:
      allOf:
      - $ref: '#/components/schemas/PartyRole'
      example:
        contactMedium:
        - '@baseType': '@baseType'
          validFor:
            startDateTime: 1985-04-12T23:20:50.52Z
            endDateTime: 1985-04-12T23:20:50.52Z
          '@type': '@type'
          contactType: contactType
          id: id
          '@schemaLocation': '@schemaLocation'
          preferred: true
        - '@baseType': '@baseType'
          validFor:
            startDateTime: 1985-04-12T23:20:50.52Z
            endDateTime: 1985-04-12T23:20:50.52Z
          '@type': '@type'
          contactType: contactType
          id: id
          '@schemaLocation': '@schemaLocation'
          preferred: true
        creditProfile:
        - creditProfileDate: 2000-01-23T04:56:07.000+00:00
          creditScore: 6
          '@baseType': '@baseType'
          validFor:
            startDateTime: 1985-04-12T23:20:50.52Z
            endDateTime: 1985-04-12T23:20:50.52Z
          '@type': '@type'
          creditRiskRating: 0
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        - creditProfileDate: 2000-01-23T04:56:07.000+00:00
          creditScore: 6
          '@baseType': '@baseType'
          validFor:
            startDateTime: 1985-04-12T23:20:50.52Z
            endDateTime: 1985-04-12T23:20:50.52Z
          '@type': '@type'
          creditRiskRating: 0
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        role: role
        agreement:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        validFor:
          startDateTime: 1985-04-12T23:20:50.52Z
          endDateTime: 1985-04-12T23:20:50.52Z
        '@type': '@type'
        description: description
        relatedParty:
        - partyOrPartyRole:
            '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            href: href
            id: id
            '@schemaLocation': '@schemaLocation'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          '@schemaLocation': '@schemaLocation'
        - partyOrPartyRole:
            '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            href: href
            id: id
            '@schemaLocation': '@schemaLocation'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          '@schemaLocation': '@schemaLocation'
        characteristic:
        - characteristicRelationship:
          - '@baseType': '@baseType'
            relationshipType: relationshipType
            '@type': '@type'
            id: id
            '@schemaLocation': '@schemaLocation'
          - '@baseType': '@baseType'
            relationshipType: relationshipType
            '@type': '@type'
            id: id
            '@schemaLocation': '@schemaLocation'
          '@baseType': '@baseType'
          '@type': '@type'
          valueType: valueType
          name: name
          id: id
          '@schemaLocation': '@schemaLocation'
        - characteristicRelationship:
          - '@baseType': '@baseType'
            relationshipType: relationshipType
            '@type': '@type'
            id: id
            '@schemaLocation': '@schemaLocation'
          - '@baseType': '@baseType'
            relationshipType: relationshipType
            '@type': '@type'
            id: id
            '@schemaLocation': '@schemaLocation'
          '@baseType': '@baseType'
          '@type': '@type'
          valueType: valueType
          name: name
          id: id
          '@schemaLocation': '@schemaLocation'
        statusReason: statusReason
        '@baseType': '@baseType'
        partyRoleSpecification:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        name: name
        paymentMethod:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        href: href
        id: id
        '@schemaLocation': '@schemaLocation'
        engagedParty:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        account:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          href: href
          id: id
          '@schemaLocation': '@schemaLocation'
        status: status
      type: object
...
    Customer_FVO:
      allOf:
      - $ref: '#/components/schemas/PartyRole_FVO'
      required:
      - engagedParty
      type: object
...
    Entity:
      allOf:
      - $ref: '#/components/schemas/Extensible'
      - $ref: '#/components/schemas/Addressable'
      description: Base entity schema for use in TMForum Open-APIs. Property.
      type: object
...
    Extensible:
      description: Base Extensible schema for use in TMForum Open-APIs - When used
        for in a schema it means that the Entity described by the schema  MUST be
        extended with the @type
      properties:
        '@type':
          description: "When sub-classing, this defines the sub-class Extensible name"
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          type: string
      required:
      - '@type'
      type: object
...
    PartyRef:
      allOf:
      - $ref: '#/components/schemas/EntityRef'
      description: A Party reference
      discriminator:
        mapping:
          PartyRef: '#/components/schemas/PartyRef'
        propertyName: '@type'
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        href: href
        id: id
        '@schemaLocation': '@schemaLocation'
      type: object
...
    PartyRole:
      allOf:
      - $ref: '#/components/schemas/Entity'
      - description: The part played by a party in a given context.
        properties:
          name:
            description: "A word, term, or phrase by which the PartyRole is known\
              \ and distinguished from other PartyRoles. It's the name of the PartyRole\
              \ unique entity."
            type: string
          description:
            description: A description of the PartyRole.
            type: string
          role:
            description: "Role played by the engagedParty in this context. As role\
              \ is defined by partyRoleSpecification, this role attribute can be used\
              \ to precise the role defined by partyRoleSpecification, or it can be\
              \ used to define the role in case there is no partyRoleSpecification."
            type: string
          engagedParty:
            $ref: '#/components/schemas/PartyRef'
          partyRoleSpecification:
            $ref: '#/components/schemas/PartyRoleSpecificationRef'
          characteristic:
            description: Describes the characteristic of a party role.
            items:
              $ref: '#/components/schemas/Characteristic'
            type: array
          account:
            description: Collection of accounts
            items:
              $ref: '#/components/schemas/AccountRef'
            type: array
          agreement:
            description: Collection of agreements
            items:
              $ref: '#/components/schemas/AgreementRef'
            type: array
          contactMedium:
            description: Collection of contact mediums
            items:
              $ref: '#/components/schemas/ContactMedium'
            type: array
          paymentMethod:
            description: Collection of payment methods
            items:
              $ref: '#/components/schemas/PaymentMethodRef'
            type: array
          creditProfile:
            description: Collection of credit profiles
            items:
              $ref: '#/components/schemas/CreditProfile'
            type: array
          relatedParty:
            description: Collection of related parties
            items:
              $ref: '#/components/schemas/RelatedPartyOrPartyRole'
            type: array
          status:
            description: Used to track the lifecycle status of the party role.
            type: string
          statusReason:
            description: "A string providing an explanation on the value of the status\
              \ lifecycle. For instance if the status is Rejected, statusReason will\
              \ provide the reason for rejection."
            type: string
          validFor:
            $ref: '#/components/schemas/TimePeriod'
        type: object
      discriminator:
        mapping:
          PartyRole: '#/components/schemas/PartyRole'
          Customer: '#/components/schemas/Customer'
          Supplier: '#/components/schemas/Supplier'
          BusinessPartner: '#/components/schemas/BusinessPartner'
          Consumer: '#/components/schemas/Consumer'
          Producer: '#/components/schemas/Producer'
        propertyName: '@type'
...
    PartyRoleSpecificationRef:
      allOf:
      - $ref: '#/components/schemas/EntityRef'
      description: Party role specification reference. A party role specification
        gives additional details on the part played by a party in a given context.
      discriminator:
        mapping:
          PartyRoleSpecificationRef: '#/components/schemas/PartyRoleSpecificationRef'
        propertyName: '@type'
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        href: href
        id: id
        '@schemaLocation': '@schemaLocation'
      type: object
...
    Supplier:
      allOf:
      - $ref: '#/components/schemas/PartyRole'
      description: 'When business partner is the Supplier '
      type: object
...
    Error:
      allOf:
      - $ref: '#/components/schemas/Extensible'
      - properties:
          code:
            description: "Application relevant detail, defined in the API or a common\
              \ list."
            type: string
          reason:
            description: Explanation of the reason for the error which can be shown
              to a client user.
            type: string
          message:
            description: More details and corrective actions related to the error
              which can be shown to a client user.
            type: string
          status:
            description: HTTP Error code extension
            type: string
          referenceError:
            description: URI of documentation describing the error.
            type: string
        required:
        - code
        - reason
        type: object
      description: "Used when an API throws an Error, typically with a HTTP error\
        \ response-code (3xx, 4xx, 5xx)"
      discriminator:
        mapping:
          Error: '#/components/schemas/Error'
        propertyName: '@type'
      example:
        reason: reason
        code: code
        '@baseType': '@baseType'
        '@type': '@type'
        '@schemaLocation': '@schemaLocation'
        message: message
        referenceError: referenceError
        status: status
...
  securitySchemes: {}